+++ src/main/java/com/hbm/inventory/fluid/Fluids.java
@@ -590,6 +590,7 @@ public class Fluids {
 
 		// ^ ^ ^ ^ ^ ^ ^ ^
 		//ADD NEW FLUIDS HERE
+		ROCKET_FUEL =			new FluidType("ROCKET_FUEL",		0x74633B, 1, 2, 0, EnumSymbol.NONE).addContainers(new CD_Canister(0x74633B)).addTraits(new FT_Flammable(300_000), LIQUID);
 
 		File folder = MainRegistry.configHbmDir;
 		File customTypes = new File(folder.getAbsolutePath() + File.separatorChar + "hbmFluidTypes.json");
@@ -695,6 +696,7 @@ public class Fluids {
 		metaOrder.add(DIESEL_CRACK);
 		metaOrder.add(DIESEL_CRACK_REFORM);
 		metaOrder.add(KEROSENE);
+		metaOrder.add(ROCKET_FUEL);
 		metaOrder.add(KEROSENE_REFORM);
 		metaOrder.add(PETROIL);
 		metaOrder.add(PETROIL_LEADED);